+++ packages/core-js/modules/esnext.async-iterator.every.js
@@ -4,7 +4,7 @@ var $every = require('../internals/async-iterator-iteration').every;
 
 // `AsyncIterator.prototype.every` method
 // https://github.com/tc39/proposal-async-iterator-helpers
-$({ target: 'AsyncIterator', proto: true, real: true }, {
+$({ target: 'AsyncIterator', proto: true, real: true, forced: true }, {
   every: function every(predicate) {
     return $every(this, predicate);
   }