+++ packages/core-js/modules/esnext.async-iterator.find.js
@@ -4,7 +4,7 @@ var $find = require('../internals/async-iterator-iteration').find;
 
 // `AsyncIterator.prototype.find` method
 // https://github.com/tc39/proposal-async-iterator-helpers
-$({ target: 'AsyncIterator', proto: true, real: true }, {
+$({ target: 'AsyncIterator', proto: true, real: true, forced: true }, {
   find: function find(predicate) {
     return $find(this, predicate);
   }