+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - `Iterator.prototype.reduce`
   - `Iterator.prototype.some`
   - `Iterator.prototype.take`
+- Fixed missing forced replacement of `AsyncIterator` helpers
 - Compat data improvements:
   - [Explicit Resource Management](https://github.com/tc39/proposal-explicit-resource-management) features disabled (again) in V8 ~ Chromium 135 and re-added in 136
   - [`RegExp.escape`](https://github.com/tc39/proposal-regex-escaping) marked as [shipped from V8 ~ Chromium 136](https://issues.chromium.org/issues/353856236#comment17)